feat(gateway): accept OTLP traces over gRPC
Caddy's built-in tracing exporter speaks OTLP/gRPC, while the gateway previously accepted only OTLP/HTTP protobuf. The dedicated TraceService listener feeds the existing mapping and bounded fanout path so both transports retain identical behavior. Incidental clippy fixes keep the affected dependency closure warning-free.

Generated-By: Gothab agent CLI (model: gpt-5.6-sol)

+++ crates/binschema-runtime/src/bitstream.rs
@@ -133,7 +133,7 @@ impl BitStreamEncoder {
             // When a byte value crosses a bit boundary, the bits are written
             // starting from the LSB of the value.
             for i in 0..8u8 {
-                let bit = ((value >> i) & 1) as u8;
+                let bit = (value >> i) & 1;
                 self.write_single_bit(bit);
             }
         }
